+++ src/tool/cpp/cppwinrt/strings/base_coroutine.h
@@ -1,34 +1,6 @@
 
 namespace winrt::impl
 {
-    inline bool is_sta() noexcept
-    {
-        int32_t aptType;
-        int32_t aptTypeQualifier;
-        return (error_ok == WINRT_CoGetApartmentType(&aptType, &aptTypeQualifier)) && ((aptType == 0 /*APTTYPE_STA*/) || (aptType == 3 /*APTTYPE_MAINSTA*/));
-    }
-
-    template <typename Async>
-    void blocking_suspend(Async const& async)
-    {
-        WINRT_ASSERT(!is_sta());
-
-        slim_mutex m;
-        slim_condition_variable cv;
-        bool completed = false;
-        async.Completed([&](auto&&...)
-        {
-            {
-                slim_lock_guard const guard(m);
-                completed = true;
-            }
-            cv.notify_one();
-        });
-
-        slim_lock_guard guard(m);
-        cv.wait(m, [&] { return completed; });
-    }
-
     template <typename Async>
     struct await_adapter
     {
@@ -320,169 +292,6 @@ namespace winrt
 
         return awaitable{ handle, timeout };
     }
-
-    struct overlapped_io
-    {
-        uintptr_t Internal;
-        uintptr_t InternalHigh;
-        union
-        {
-            struct
-            {
-                uint32_t Offset;
-                uint32_t OffsetHigh;
-            } s;
-            void* Pointer;
-        };
-        void* hEvent;
-    };
-
-    struct awaitable_base
-    {
-        static void WINRT_CALL callback(void*, void*, void* overlapped, uint32_t result, std::size_t, void*) noexcept
-        {
-            auto context = static_cast<awaitable_base*>(overlapped);
-            context->m_result = result;
-            context->m_resume();
-        }
-
-    protected:
-
-        overlapped_io m_overlapped{};
-        uint32_t m_result{};
-        std::experimental::coroutine_handle<> m_resume{ nullptr };
-    };
-
-    struct resumable_io
-    {
-        resumable_io(void* object) :
-            m_io(check_pointer(WINRT_CreateThreadpoolIo(object, awaitable_base::callback, nullptr, nullptr)))
-        {
-        }
-
-        template <typename F>
-        auto start(F callback)
-        {
-            struct awaitable : awaitable_base, F
-            {
-                awaitable(impl::ptp_io io, F callback) noexcept :
-                    F(callback),
-                    m_io(io)
-                {}
-
-                bool await_ready() const noexcept
-                {
-                    return false;
-                }
-
-                void await_suspend(std::experimental::coroutine_handle<> resume_handle)
-                {
-                    m_resume = resume_handle;
-                    WINRT_StartThreadpoolIo(m_io);
-
-                    try
-                    {
-                        (*this)(m_overlapped);
-                    }
-                    catch (...)
-                    {
-                        WINRT_CancelThreadpoolIo(m_io);
-                        throw;
-                    }
-                }
-
-                uint32_t await_resume() const
-                {
-                    if (m_result != 38 /*ERROR_HANDLE_EOF*/)
-                    {
-                        check_win32(m_result);
-                    }
-                    return static_cast<uint32_t>(m_overlapped.InternalHigh);
-                }
-
-                impl::ptp_io m_io{};
-            };
-
-            return awaitable(get(), callback);
-        }
-
-        template <typename F>
-        auto start_pending(F callback)
-        {
-            struct awaitable : awaitable_base, F
-            {
-                awaitable(impl::ptp_io io, F callback) noexcept :
-                    F(callback),
-                    m_io(io)
-                {}
-
-                bool await_ready() const noexcept
-                {
-                    return false;
-                }
-
-                bool await_suspend(std::experimental::coroutine_handle<> resume_handle)
-                {
-                    m_resume = resume_handle;
-                    WINRT_StartThreadpoolIo(m_io);
-
-                    try
-                    {
-                        bool const pending = (*this)(m_overlapped);
-
-                        if (!pending)
-                        {
-                            WINRT_CancelThreadpoolIo(m_io);
-                        }
-
-                        return pending;
-                    }
-                    catch (...)
-                    {
-                        WINRT_CancelThreadpoolIo(m_io);
-                        throw;
-                    }
-                }
-
-                uint32_t await_resume() const
-                {
-                    if (m_result != 38 /*ERROR_HANDLE_EOF*/)
-                    {
-                        check_win32(m_result);
-                    }
-                    return static_cast<uint32_t>(m_overlapped.InternalHigh);
-                }
-
-                impl::ptp_io m_io{};
-            };
-
-            return awaitable(get(), callback);
-        }
-
-        impl::ptp_io get() const noexcept
-        {
-            return m_io.get();
-        }
-
-    private:
-
-        struct io_traits
-        {
-            using type = impl::ptp_io;
-
-            static void close(type value) noexcept
-            {
-                WINRT_CloseThreadpoolIo(value);
-            }
-
-            static constexpr type invalid() noexcept
-            {
-                return nullptr;
-            }
-        };
-
-        handle_type<io_traits> m_io;
-    };
 
 #ifdef _RESUMABLE_FUNCTIONS_SUPPORTED
     inline auto operator co_await(Windows::Foundation::TimeSpan duration)
