+++ src/tool/cpp/cppwinrt/strings/base_dependencies.h
@@ -20,7 +20,6 @@
 #include <utility>
 #include <unordered_map>
 #include <vector>
-#include <experimental/coroutine>
 
 #if __has_include(<WindowsNumerics.impl.h>)
 #define WINRT_NUMERICS
